🚜 Rename : pages폴더로 이름 변경 및 GlobalStyles와 router 분리
